Updated: Provider names for Beam and Hitbox.
