+++ src/providers/index.ts
@@ -1,5 +1,5 @@
-export * from './beam';
-export * from './hitbox';
 export * from './patreon';
+export * from './mixer';
 export * from './picarto';
+export * from './smashcast';
 export * from './twitch';
